change product property type

+++ src/backend/Domain/Entites/Products/Product.cs
@@ -9,9 +9,9 @@ namespace Domain.Entites
         public int CategoryId { get; set; }
         public string QuantityPerUnit { get; set; }
         public decimal UnitPrice { get; set; }
-        public int UnitsInStock { get; set; }
-        public int UnitsOnOrder { get; set; }
-        public int ReorderLevel { get; set; }
+        public short UnitsInStock { get; set; }
+        public short UnitsOnOrder { get; set; }
+        public short ReorderLevel { get; set; }
         public bool Discontinued { get; set; }
 
         public Supplier Supplier { get; set; }
