+++ src/backend/Domain/Models/Products/ProductModel.cs
@@ -10,9 +10,9 @@ namespace Domain.Models
         public int CategoryId { get; set; }
         public string QuantityPerUnit { get; set; }
         public decimal UnitPrice { get; set; }
-        public int UnitsInStock { get; set; }
-        public int UnitsOnOrder { get; set; }
-        public int ReorderLevel { get; set; }
+        public short UnitsInStock { get; set; }
+        public short UnitsOnOrder { get; set; }
+        public short ReorderLevel { get; set; }
         public bool Discontinued { get; set; }
 
         [JsonIgnore]
