+++ src/backend/Service/Valiadations/FluentValidation/Products/ProductModelValidator.cs
@@ -10,7 +10,7 @@ namespace Service.Valiadations.FluentValidation
         {
             RuleFor(x => x.Name).NotNull().WithMessage("{PropertyName} is reqired").NotEmpty().WithMessage("{PropertyName} is reqired");
             RuleFor(x => x.UnitPrice).InclusiveBetween(1, decimal.MaxValue).WithMessage("{PropertyName} must be greater 0");
-            RuleFor(x => x.UnitsInStock).InclusiveBetween(1, int.MaxValue).WithMessage("{PropertyName} must be greater 0");
+            RuleFor(x => (int)x.UnitsInStock).InclusiveBetween(1, short.MaxValue).WithMessage("{PropertyName} must be greater 0");
             RuleFor(x => x.CategoryId).InclusiveBetween(1, int.MaxValue).WithMessage("{PropertyName} must be greater 0");
         }
     }
